Add support for device tree updating
NOTE: This fix is temporary and will be modifed once
https://github.com/advancedtelematic/meta-updater/pull/728 will be
backported to the supported branches, allowing to use the variable
OSTREE_DEVICETREE instead of overwriting KERNEL_DEVICETREE.

This is necessary when the kernel and device tree are updated.

OSTREE_DEPLOY_DEVICETREE is a meta-updater variable that will include
the device tree to the update.

KERNEL_DEVICETREE is overwritten in fullmetalupdate-os-package's
environment because there are multiple device tree that are defined in
the machine conf, and OSTree is not able to determine the correct device
tree to select on its own. Replacing this variable makes sure only one
device tree (the correct one) is shipped.

bootcmd_load_fdt loads the new device tree.
bootcmd_cond has changed to take the changes into account.